+++ src/main/resources/liquibase/scripts/script.sql
@@ -3,18 +3,18 @@
 -- changeset Karachevtsev:1
 CREATE TABLE мониторы (
     id SERIAL PRIMARY KEY,
-    серийный_номер INTEGER,
+    серийный_номер BIGINT,
     производитель VARCHAR(255),
-    цена DECIMAL(10, 2),
+    цена INT,
     количество INT,
     диагональ INT
 );
 
 CREATE TABLE жесткие_диски (
     id SERIAL PRIMARY KEY,
-    серийный_номер INTEGER,
+    серийный_номер BIGINT,
     производитель VARCHAR(255),
-    цена DECIMAL(10, 2),
+    цена INT,
     количество INT,
     память INT
 );
